backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
cffi==2.0.0
click==8.3.2
cryptography==46.0.7
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.136.0
h11==0.16.0
httptools==0.7.1
idna==3.11
pycparser==3.0
pydantic==2.13.2
pydantic_core==2.46.2
PyMySQL==1.1.2
python-dotenv==1.2.2
PyYAML==6.0.3
SQLAlchemy==2.0.49
sqlmodel==0.0.38
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.44.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
